+++ embed-pdf-viewer.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name:       Embed PDF Viewer
  * Plugin URI:        https://github.com/afragen/embed-pdf-viewer
- * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed into an `<object>` tag or Google Doc Viewer as fallback.
+ * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed into an `object` tag or Google Doc Viewer as fallback.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
  * Version:           1.4.0