+++ embed-pdf-viewer.php
@@ -5,7 +5,7 @@
  * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed into an `object` tag or Google Doc Viewer as fallback.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           1.4.0
+ * Version:           1.5.0
  * License:           GPLv2+
  * Domain Path:       /languages
  * Text Domain:       embed-pdf-viewer
@@ -135,18 +135,34 @@ class Embed_PDF_Viewer {
 			$atts['title'] = ucwords( preg_replace( '/(-|_)/', ' ', $post->post_name ) );
 		}
 
-		$iframe_fallback = '<iframe src="https://docs.google.com/viewer?url=' . urlencode( $post->guid );
+		$iframe_fallback = '<iframe class="embed-pdf-viewer" src="https://docs.google.com/viewer?url=' . urlencode( $post->guid );
 		$iframe_fallback .= '&amp;embedded=true" frameborder="0" ';
 		$iframe_fallback .= 'style="height:' . $atts['height'] . 'px;width:' . $atts['width'] . 'px;" ';
 		$iframe_fallback .= 'title="' . $atts['title'] . '"></iframe>' . "\n";
 
-		$embed = '<object data="' . $post->guid;
-		$embed .= '#scrollbar=1&toolbar=1"';
-		$embed .= 'type="application/pdf" ';
-		$embed .= 'height=' . $atts['height'] . ' width=' . $atts['width'] . ' > ';
-		$embed .= $iframe_fallback;
-		$embed .= '</object>';
+		/*
+		 * Because <object> doesn't scroll in iOS
+		 * Hide <object> in iOS and hide <iframe> in not iOS
+		 */
+		$style = '<style>
+		@media only screen and (max-device-width: 1024px) {
+			object.embed-pdf-viewer { display:none; }			
+		}
+		@media only screen and (min-device-width : 1024px) {
+			iframe.embed-pdf-viewer { display:none; }
+			object iframe.embed-pdf-viewer { display:block; }
+		}
+		</style>';
 
+		$object = '<object class="embed-pdf-viewer" data="' . $post->guid;
+		$object .= '#scrollbar=1&toolbar=1"';
+		$object .= 'type="application/pdf" ';
+		$object .= 'height=' . $atts['height'] . ' width=' . $atts['width'] . ' > ';
+		$object .= $iframe_fallback;
+		$object .= '</object>';
+
+		$embed = $object;
+		$embed .= $style . $iframe_fallback;
 		$embed .= '<a href="' . $post->guid . '">' . $atts['title'] . '</a>';
 
 		return $embed;