+++ CHANGES.md
@@ -1,7 +1,7 @@
 #### [unreleased]
 
 #### 1.4.0 / 2016-12-16
-* switch to using `<object>`, with `<iframe>` as fallback display method
+* switch to using `object` tag, with `iframe` as fallback display method
 
 #### 1.3.0 / 2016-08-13
 * bugfix for not returning non-PDF file from Media Library